pyproject.toml
[project]
name = "mycqu-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Zhu Zijun", email = "2961163526@qq.com" }
]
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.69.0",
    "httpx>=0.28.1",
    "mycqu",
    "protobuf==4.24.2",
    "pydantic>=2.10.5",
    "tencentserverless>=0.1.7",
]

[tool.uv.sources]
mycqu = { path = "../pymycqu", editable = true }

#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"
